consumer/requirements.txt
kafka-python==1.4.6
python-redis==0.2.2
